requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.2.0
baostock==0.8.8
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.4
cx-Oracle==8.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
jsonpath==0.82
kafka-python==2.0.2
numpy==1.24.1
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
paho-mqtt==1.6.1
pandas==1.5.2
paramiko==2.12.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==22.1.0
PySocks==1.7.1
pytest==7.2.0
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
